renamed optimizations, added logging, documented, positivity is the main optimization now
