+++ man/optimize_coordinate_descent.Rd
@@ -1,18 +1,18 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RcppExports.R
-\name{theta_derivative_stage2}
-\alias{theta_derivative_stage2}
-\title{Main function to calculate error terms}
+\name{optimize_coordinate_descent}
+\alias{optimize_coordinate_descent}
+\title{Main training loop with all gradient steps. This is naive algorithm from version 1.0.
+It optimizes 3 terms (5 in experimental version). 
+Gradient steps performed in X and Omega spaces separatelly with NNLS used to merge them together into deconvolution term.}
 \usage{
-theta_derivative_stage2(
+optimize_coordinate_descent(
   X,
   Omega,
   D_w,
   SVRt,
   R,
   S,
-  X_center,
-  Omega_center,
   coef_der_X,
   coef_der_Omega,
   coef_hinge_H,
@@ -28,7 +28,9 @@ theta_derivative_stage2(
   r_const_X = 0,
   r_const_Omega = 0,
   thresh = 0.8,
-  theta_threshold = 0
+  convergence_tol = 1e-12,
+  stop_criteria_window = 100000L,
+  debug_stats = FALSE
 )
 }
 \arguments{
@@ -43,10 +45,6 @@ theta_derivative_stage2(
 \item{R}{current R}
 
 \item{S}{current S}
-
-\item{X_center}{optimization restriction directions for X.}
-
-\item{Omega_center}{optimization restriction direcitons of Omega.}
 
 \item{coef_der_X}{learning rate X}
 
@@ -78,11 +76,17 @@ theta_derivative_stage2(
 
 \item{thresh}{experimental. not tested}
 
-\item{theta_threshold}{angle to restrict optimization/}
+\item{convergence_tol}{tolerance for convergence.}
+
+\item{stop_criteria_window}{how long error should be on plateu to decrease the learning rate}
+
+\item{debug_stats}{wether to save grad norm values.}
 }
 \value{
 new parameters
 }
 \description{
-Main function to calculate error terms
+Main training loop with all gradient steps. This is naive algorithm from version 1.0.
+It optimizes 3 terms (5 in experimental version). 
+Gradient steps performed in X and Omega spaces separatelly with NNLS used to merge them together into deconvolution term.
 }
