+++ man/optimize_positivity.Rd
@@ -1,10 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RcppExports.R
-\name{alternative_derivative_stage2}
-\alias{alternative_derivative_stage2}
-\title{Main function to calculate error terms}
+\name{optimize_positivity}
+\alias{optimize_positivity}
+\title{Main training loop with all gradient steps. This is the main algorithm for now. 
+It optimizes two positivity terms ensuring invertability of both X and Omega. 
+Gradient steps performed in X space.}
 \usage{
-alternative_derivative_stage2(
+optimize_positivity(
   X,
   Omega,
   D_w,
@@ -69,5 +71,7 @@ alternative_derivative_stage2(
 new parameters
 }
 \description{
-Main function to calculate error terms
+Main training loop with all gradient steps. This is the main algorithm for now. 
+It optimizes two positivity terms ensuring invertability of both X and Omega. 
+Gradient steps performed in X space.
 }
