+++ man/optimize_theta.Rd
@@ -1,16 +1,18 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/RcppExports.R
-\name{derivative_stage2}
-\alias{derivative_stage2}
-\title{Main function to calculate error terms}
+\name{optimize_theta}
+\alias{optimize_theta}
+\title{Main training loop with all gradient steps. This implementation tries to move points within the specified angle theta only.}
 \usage{
-derivative_stage2(
+optimize_theta(
   X,
   Omega,
   D_w,
   SVRt,
   R,
   S,
+  X_center,
+  Omega_center,
   coef_der_X,
   coef_der_Omega,
   coef_hinge_H,
@@ -26,9 +28,7 @@ derivative_stage2(
   r_const_X = 0,
   r_const_Omega = 0,
   thresh = 0.8,
-  convergence_tol = 1e-12,
-  stop_criteria_window = 100000L,
-  debug_stats = FALSE
+  theta_threshold = 0
 )
 }
 \arguments{
@@ -43,6 +43,10 @@ derivative_stage2(
 \item{R}{current R}
 
 \item{S}{current S}
+
+\item{X_center}{optimization restriction directions for X.}
+
+\item{Omega_center}{optimization restriction direcitons of Omega.}
 
 \item{coef_der_X}{learning rate X}
 
@@ -74,15 +78,11 @@ derivative_stage2(
 
 \item{thresh}{experimental. not tested}
 
-\item{convergence_tol}{tolerance for convergence.}
-
-\item{stop_criteria_window}{how long error should be on plateu to decrease the learning rate}
-
-\item{debug_stats}{wether to save grad norm values.}
+\item{theta_threshold}{angle to restrict optimization/}
 }
 \value{
 new parameters
 }
 \description{
-Main function to calculate error terms
+Main training loop with all gradient steps. This implementation tries to move points within the specified angle theta only.
 }
